Got all tests passing on local test server.
Tests now pass. This was accomplished by making sure the productInfoCorrect
test was lead to the correct page to test product info as it seems the page
after clicking submit for a new product has changed. Also made selects more
robust. Before which items to select had to be hard coded which may have resulted
in false failures because sometimes the number of select elements change. To fix
this I've added a new method SelectFirstNonEmptyElement.