Fixing a few broken tests
There have been a few UI changes so some tests needed to be changed.
Mainly, pnc doesn't allow multiple projects/products to be made anymore.
This is good behaviour but the problem is that some of the tests now
fail because the test server retains previous test products, when the
webdriver tries to add another it throws an error.