Use random numbers for item names
Because pnc e2e tests on the autodeploy server, the items it adds
are persistent. So if the tests use the same name for products, configurations
and etc. then the tests will only pass the first time (because it will try to
make a duplicate which will fail causing the test to fail). This has been *fixed*
by using random numbers for item names instead. Of course this could still fail,
but there's only a 1/100000 chance.