Delegating WebDriver setup to new Task classes
UITester is becoming something of a god class. It basically does
all of the webdriving related stuff. I've added a new task class.
The plan is that tasks will simply take a webdriver and perform
all the actions. This is so that as new tasks are added they can
be added and organized into packages instead of creating a massive
single class (UITester).