Adding assert to addBuildConfiguration, cleanups
Making sure title is set properly when adding build to a
BuildConfigurationSet. Ignoring ComponentBuildTest, changed assert
in BuildRecordTest to make it search for a running build and not
instead of a link to the build configuration. This way the test
should not get confused when the build has been completed before
but is currently running.